# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Space Caps Store API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Product {
  id: number;
  name: string;
  brand: string;
  price: number;
  imageUrl: string;
  /** @nullable */
  imageBackUrl?: string | null;
  /** @nullable */
  description?: string | null;
  available: boolean;
  featured?: boolean;
}

export interface Brand {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  logoUrl?: string | null;
}

export interface CartItem {
  id: number;
  productId: number;
  productName: string;
  productBrand: string;
  price: number;
  quantity: number;
  imageUrl: string;
}

export interface Cart {
  sessionId: string;
  items: CartItem[];
  total: number;
}

export interface CartItemInput {
  sessionId: string;
  productId: number;
  quantity: number;
}

export type ListProductsParams = {
  brand?: string;
  search?: string;
};

export type GetCartParams = {
  sessionId: string;
};