# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Space Caps Store API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: products
    description: Cap product operations
  - name: cart
    description: Shopping cart operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List all products
      parameters:
        - name: brand
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get a single product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found

  /brands:
    get:
      operationId: listBrands
      tags: [products]
      summary: List all brands
      responses:
        "200":
          description: List of brands
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Brand"

  /products/featured:
    get:
      operationId: getFeaturedProducts
      tags: [products]
      summary: Get featured/hero products
      responses:
        "200":
          description: Featured products for homepage hero
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"

  /cart:
    get:
      operationId: getCart
      tags: [cart]
      summary: Get cart contents
      parameters:
        - name: sessionId
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cart contents
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"
    post:
      operationId: addToCart
      tags: [cart]
      summary: Add item to cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CartItemInput"
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

  /cart/{itemId}:
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Remove item from cart
      parameters:
        - name: itemId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Cart"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    Product:
      type: object
      required: [id, name, brand, price, imageUrl, available]
      properties:
        id:
          type: integer
        name:
          type: string
        brand:
          type: string
        price:
          type: number
        imageUrl:
          type: string
        imageBackUrl:
          type: ["string", "null"]
        description:
          type: ["string", "null"]
        available:
          type: boolean
        featured:
          type: boolean

    Brand:
      type: object
      required: [id, name, slug]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        logoUrl:
          type: ["string", "null"]

    Cart:
      type: object
      required: [sessionId, items, total]
      properties:
        sessionId:
          type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/CartItem"
        total:
          type: number

    CartItem:
      type: object
      required: [id, productId, productName, productBrand, price, quantity, imageUrl]
      properties:
        id:
          type: integer
        productId:
          type: integer
        productName:
          type: string
        productBrand:
          type: string
        price:
          type: number
        quantity:
          type: integer
        imageUrl:
          type: string

    CartItemInput:
      type: object
      required: [sessionId, productId, quantity]
      properties:
        sessionId:
          type: string
        productId:
          type: integer
        quantity:
          type: integer