# lib/api-zod/src/generated/types/brand.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Space Caps Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Brand {
  id: number;
  name: string;
  slug: string;
  /** @nullable */
  logoUrl?: string | null;
}