# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Space Caps Store API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  brand: string;
  price: number;
  imageUrl: string;
  /** @nullable */
  imageBackUrl?: string | null;
  /** @nullable */
  description?: string | null;
  available: boolean;
  featured?: boolean;
}